\docType{package}
\name{sumFREGAT-package}
\alias{sumFREGAT-package}
\title{sumFREGAT: Fast REGional Association Tests on summary statistics}
\description{
	The sumFREGAT package computes the most popular and efficient
	tests for the region-based association analysis using summary statistics data
	(beta and P values) and correlations between genetic variants. It does not
	require genotype or phenotype data. Methods implemented are SKAT and SKAT-O
	(sequence kernel association tests), BT (burden test), ACAT (aggregated Cauchy
	association test), FLM (functional linear model), PCA (principal components analysis),
	and others.
}
\details{
Package: sumFREGAT\cr
Type: Package\cr
License: GPLv3
}
\author{
	Nadezhda Belonogova <belon@bionet.nsc.ru>\cr
	Gulnara Svishcheva <gulsvi@bionet.nsc.ru>\cr\cr
	The authors are grateful to Dr. Anatoly Kirichenko for technical support\cr
	and Prof. Tatiana Axenovich for scientific guidance.\cr
}
